+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `ae.AllStorageSOPClasses`: pynetdicom `_STORAGE_CLASSES` catalog for Storage SCP abstract syntaxes
+- `AcceptedAbstractSyntaxes` may include `"*"` to accept any peer-proposed Abstract Syntax (opt-in escape hatch)
+
 ## [0.9.0] - 2026-07-14
 
 ### Added